0.2.0

+++ src/cooldown.rs
@@ -3,8 +3,8 @@ use std::ops::Sub;
 use std::time::{Duration, Instant};
 
 pub struct CoolDown {
-    pub available: HashMap<u8, Duration>,
-    pub current: u8,
+    pub available: HashMap<u32, Duration>,
+    pub current: u32,
     pub ready_in: u128,
     pub completion_ratio: f32,
     pub last_used: Instant,
